feat: add currentNav utils
